users can no longer check meh profile

+++ cogs/anti_magik.py
@@ -14,7 +14,21 @@ class Owner:
 			def pred(m):
 				return m.author.id == 270904126974590976 and m.channel == m.channel
 			try:
-				msg = await self.bot.wait_for('message', check=pred, timeout=10.0)
+				msg = await self.bot.wait_for('message', check=pred, timeout=120.0)
+			except asyncio.TimeoutError:
+				async for i in m.channel.history(limit=5):
+					await i.delete()
+			else:
+				await asyncio.sleep(0.5)
+				await msg.delete()
+				await m.channel.send("next time i ban you")
+		commands = ["t!avatar <@264838866480005122>", ".avatar <@264838866480005122>", "./avatar <@264838866480005122>"]
+		bots = [506735111543193601, 418412306981191680, 172002275412279296]
+		if m.content.lower() in commands:
+			def pred(m):
+				return m.author.id in bots and m.channel == m.channel
+			try:
+				msg = await self.bot.wait_for('message', check=pred, timeout=60.0)
 			except asyncio.TimeoutError:
 				async for i in m.channel.history(limit=5):
 					await i.delete()
